Revert "refactor: header options v2 styling into hook"
This reverts commit ed78e969eeb04084f5bb83a8c849d0b8cd0f65f2.
