+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/PortfolioNavigator.tsx
@@ -2,7 +2,7 @@ import { NavigationProp, useNavigation } from '@react-navigation/native'
 import { createStackNavigator } from '@react-navigation/stack'
 import { WalletToken } from '@store/wallet'
 import BigNumber from 'bignumber.js'
-import { Image, TouchableOpacity, View } from 'react-native'
+import { TouchableOpacity, View } from 'react-native'
 import { BarCodeScanner } from '@components/BarCodeScanner'
 import { ConnectionStatus, HeaderTitle } from '@components/HeaderTitle'
 import { getNativeIcon } from '@components/icons/assets'
@@ -43,10 +43,6 @@ import { TransactionDetailScreen } from '@screens/AppNavigator/screens/Transacti
 import { VMTransaction } from '@screens/AppNavigator/screens/Transactions/screens/stateProcessor'
 import { HeaderNetworkStatus } from '@components/HeaderNetworkStatus'
 import { useNavigatorScreenOptions } from '@hooks/useNavigatorScreenOptions'
-import { useThemeContext } from '@shared-contexts/ThemeProvider'
-import GridBackgroundImageLight from '@assets/images/onboarding/grid-background-light.png'
-import GridBackgroundImageDark from '@assets/images/onboarding/grid-background-dark.png'
-import { useHeaderScreenOptionsV2 } from '@hooks/useHeaderScreenOptionsV2'
 
 export interface PortfolioParamList {
   PortfolioScreen: undefined
@@ -128,41 +124,30 @@ export function PortfolioNavigator (): JSX.Element {
   const navigation = useNavigation<NavigationProp<PortfolioParamList>>()
   const headerContainerTestId = 'portfolio_header_container'
   const { isFeatureAvailable } = useFeatureFlagContext()
-  const { isLight } = useThemeContext()
 
   const goToNetworkSelect = (): void => {
     navigation.navigate('NetworkDetails')
   }
-  const HeaderScreenOptionsV2 = useHeaderScreenOptionsV2({ headerNetworkOnPress: goToNetworkSelect })
   const screenOptions = useNavigatorScreenOptions()
-
-  function PortfolioStacks (): JSX.Element {
-    return (
-      <PortfolioStack.Navigator
-        initialRouteName='PortfolioScreen'
-        screenOptions={{
+  return (
+    <PortfolioStack.Navigator
+      initialRouteName='PortfolioScreen'
+      screenOptions={{
         headerTitleAlign: 'center'
       }}
-      >
-        <PortfolioStack.Screen
-          component={isFeatureAvailable('setting_v2') ? SettingsNavigatorV2 : SettingsNavigator}
-          name={translate('PortfolioNavigator', 'Settings')}
-          options={{
+    >
+      <PortfolioStack.Screen
+        component={isFeatureAvailable('setting_v2') ? SettingsNavigatorV2 : SettingsNavigator}
+        name={translate('PortfolioNavigator', 'Settings')}
+        options={{
           headerShown: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={PortfolioScreen}
-          name='PortfolioScreen'
-          options={{
-          headerBackground: () => (
-            <Image
-              source={isLight ? GridBackgroundImageLight : GridBackgroundImageDark}
-              style={{ height: 220, width: '100%' }}
-              resizeMode='cover'
-            />
-          ),
+      <PortfolioStack.Screen
+        component={PortfolioScreen}
+        name='PortfolioScreen'
+        options={{
           headerLeft: () => (
             <TouchableOpacity
               onPress={() => navigation.navigate('Settings')}
@@ -186,12 +171,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={ReceiveScreen}
-          name='Receive'
-          options={{
+      <PortfolioStack.Screen
+        component={ReceiveScreen}
+        name='Receive'
+        options={{
           ...screenOptions,
           headerRight: () => (
             <HeaderNetworkStatus onPress={goToNetworkSelect} />
@@ -199,12 +184,12 @@ export function PortfolioNavigator (): JSX.Element {
           headerBackTitleVisible: false,
           headerTitle: translate('screens/ReceiveScreen', 'Receive')
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={AddressControlScreen}
-          name='AddressControlScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={AddressControlScreen}
+        name='AddressControlScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/AddressControlScreen', 'Wallet Address')}
@@ -214,12 +199,12 @@ export function PortfolioNavigator (): JSX.Element {
           headerRightContainerStyle: tailwind('px-2 py-2'),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={GetDFIScreen}
-          name='GetDFIScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={GetDFIScreen}
+        name='GetDFIScreen'
+        options={{
           ...screenOptions,
           headerRight: () => (
             <HeaderNetworkStatus onPress={goToNetworkSelect} />
@@ -227,12 +212,12 @@ export function PortfolioNavigator (): JSX.Element {
           headerBackTitleVisible: false,
           headerTitle: translate('screens/ReceiveScreen', 'Get DFI')
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={MarketplaceScreen}
-          name='MarketplaceScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={MarketplaceScreen}
+        name='MarketplaceScreen'
+        options={{
           ...screenOptions,
           headerRight: () => (
             <HeaderNetworkStatus onPress={goToNetworkSelect} />
@@ -240,12 +225,12 @@ export function PortfolioNavigator (): JSX.Element {
           headerBackTitleVisible: false,
           headerTitle: translate('screens/MarketplaceScreen', 'Marketplace')
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={SendScreen}
-          name='Send'
-          options={{
+      <PortfolioStack.Screen
+        component={SendScreen}
+        name='Send'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/SendScreen', 'Send')}
@@ -254,12 +239,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={SendConfirmationScreen}
-          name='SendConfirmationScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={SendConfirmationScreen}
+        name='SendConfirmationScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/SendConfirmationScreen', 'Confirm Send')}
@@ -268,12 +253,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={TokenDetailScreen}
-          name='TokenDetail'
-          options={({ route }: { route: any }) => ({
+      <PortfolioStack.Screen
+        component={TokenDetailScreen}
+        name='TokenDetail'
+        options={({ route }: { route: any }) => ({
           headerTitle: () => {
             const token = route?.params?.token
             const Icon = getNativeIcon(token.displaySymbol)
@@ -295,12 +280,12 @@ export function PortfolioNavigator (): JSX.Element {
           },
           headerBackTitleVisible: false
         })}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={ConvertScreen}
-          name='Convert'
-          options={{
+      <PortfolioStack.Screen
+        component={ConvertScreen}
+        name='Convert'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/ConvertScreen', 'Convert DFI')}
@@ -309,12 +294,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={ConvertConfirmationScreen}
-          name='ConvertConfirmationScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={ConvertConfirmationScreen}
+        name='ConvertConfirmationScreen'
+        options={{
           headerBackTitleVisible: false,
           headerTitle: () => (
             <HeaderTitle
@@ -323,12 +308,12 @@ export function PortfolioNavigator (): JSX.Element {
             />
           )
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={BarCodeScanner}
-          name='BarCodeScanner'
-          options={{
+      <PortfolioStack.Screen
+        component={BarCodeScanner}
+        name='BarCodeScanner'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/ConvertScreen', 'Scan recipient QR')}
@@ -337,12 +322,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={TokensVsUtxoScreen}
-          name='TokensVsUtxo'
-          options={{
+      <PortfolioStack.Screen
+        component={TokensVsUtxoScreen}
+        name='TokensVsUtxo'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/ConvertScreen', 'UTXO vs Token')}
@@ -351,22 +336,22 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={NetworkDetails}
-          name='NetworkDetails'
-          options={{
+      <PortfolioStack.Screen
+        component={NetworkDetails}
+        name='NetworkDetails'
+        options={{
           headerTitle: translate('screens/NetworkDetails', 'Wallet Network'),
           headerBackTitleVisible: false,
           headerBackTestID: 'network_details_header_back'
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={AboutScreen}
-          name='AboutScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={AboutScreen}
+        name='AboutScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/AboutScreen', 'About')}
@@ -375,12 +360,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={CompositeSwapScreen}
-          name='CompositeSwap'
-          options={{
+      <PortfolioStack.Screen
+        component={CompositeSwapScreen}
+        name='CompositeSwap'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/DexScreen', 'Swap tokens')}
@@ -389,12 +374,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={ConfirmCompositeSwapScreen}
-          name='ConfirmCompositeSwapScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={ConfirmCompositeSwapScreen}
+        name='ConfirmCompositeSwapScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/DexScreen', 'Confirm swap')}
@@ -403,23 +388,23 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={AddressBookScreen}
-          name='AddressBookScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={AddressBookScreen}
+        name='AddressBookScreen'
+        options={{
           headerBackTitleVisible: false,
           headerTitle: () => (
             <></>
           )
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={AddOrEditAddressBookScreen}
-          name='AddOrEditAddressBookScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={AddOrEditAddressBookScreen}
+        name='AddOrEditAddressBookScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/AddOrEditAddressBookScreen', 'Add New Address')}
@@ -428,12 +413,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={AddLiquidityScreen}
-          name='AddLiquidity'
-          options={{
+      <PortfolioStack.Screen
+        component={AddLiquidityScreen}
+        name='AddLiquidity'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/DexScreen', 'Add Liquidity')}
@@ -442,12 +427,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={ConfirmAddLiquidityScreen}
-          name='ConfirmAddLiquidity'
-          options={{
+      <PortfolioStack.Screen
+        component={ConfirmAddLiquidityScreen}
+        name='ConfirmAddLiquidity'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/DexScreen', 'Confirm Add Liquidity')}
@@ -455,12 +440,12 @@ export function PortfolioNavigator (): JSX.Element {
             />
           )
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={WithdrawFutureSwapScreen}
-          name='WithdrawFutureSwapScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={WithdrawFutureSwapScreen}
+        name='WithdrawFutureSwapScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/WithdrawFutureSwapScreen', 'Withdraw from future swap')}
@@ -469,12 +454,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={RemoveLiquidityScreen}
-          name='RemoveLiquidity'
-          options={{
+      <PortfolioStack.Screen
+        component={RemoveLiquidityScreen}
+        name='RemoveLiquidity'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/DexScreen', 'Remove Liquidity')}
@@ -482,12 +467,12 @@ export function PortfolioNavigator (): JSX.Element {
             />
           )
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={FutureSwapScreen}
-          name='FutureSwapScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={FutureSwapScreen}
+        name='FutureSwapScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/FutureSwapScreen', 'Future Swap')}
@@ -496,12 +481,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={ConfirmWithdrawFutureSwapScreen}
-          name='ConfirmWithdrawFutureSwapScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={ConfirmWithdrawFutureSwapScreen}
+        name='ConfirmWithdrawFutureSwapScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/ConfirmWithdrawFutureSwapScreen', 'Confirm withdrawal')}
@@ -510,12 +495,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={RemoveLiquidityConfirmScreen}
-          name='RemoveLiquidityConfirmScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={RemoveLiquidityConfirmScreen}
+        name='RemoveLiquidityConfirmScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/DexScreen', 'Confirm Removal')}
@@ -523,12 +508,12 @@ export function PortfolioNavigator (): JSX.Element {
             />
           )
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={TransactionsScreen}
-          name='TransactionsScreen'
-          options={{
+      <PortfolioStack.Screen
+        component={TransactionsScreen}
+        name='TransactionsScreen'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/TransactionsScreen', 'Transactions')}
@@ -537,12 +522,12 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
+      />
 
-        <PortfolioStack.Screen
-          component={TransactionDetailScreen}
-          name='TransactionDetail'
-          options={{
+      <PortfolioStack.Screen
+        component={TransactionDetailScreen}
+        name='TransactionDetail'
+        options={{
           headerTitle: () => (
             <HeaderTitle
               text={translate('screens/TransactionDetailScreen', 'Transaction')}
@@ -551,434 +536,7 @@ export function PortfolioNavigator (): JSX.Element {
           ),
           headerBackTitleVisible: false
         }}
-        />
-      </PortfolioStack.Navigator>
-    )
-  }
-
-  function PortfolioStacksV2 (): JSX.Element {
-    return (
-      <PortfolioStack.Navigator
-        initialRouteName='PortfolioScreen'
-        screenOptions={HeaderScreenOptionsV2}
-      >
-        <PortfolioStack.Screen
-          component={isFeatureAvailable('setting_v2') ? SettingsNavigatorV2 : SettingsNavigator}
-          name={translate('PortfolioNavigator', 'Settings')}
-          options={{
-          headerShown: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={PortfolioScreen}
-          name='PortfolioScreen'
-          options={{
-          headerBackground: () => (
-            <Image
-              source={isLight ? GridBackgroundImageLight : GridBackgroundImageDark}
-              style={{ height: 220, width: '100%' }}
-              resizeMode='cover'
-            />
-          ),
-          headerLeft: () => (
-            <TouchableOpacity
-              onPress={() => navigation.navigate('Settings')}
-              testID='header_settings'
-            >
-              <ThemedIcon
-                iconType='MaterialIcons'
-                name='settings'
-                size={28}
-                style={tailwind('ml-2')}
-                light={tailwind('text-primary-500')}
-                dark={tailwind('text-darkprimary-500')}
-              />
-            </TouchableOpacity>
-          ),
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/PortfolioScreen', 'Portfolio')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={ReceiveScreen}
-          name='Receive'
-          options={{
-          ...screenOptions,
-          headerRight: () => (
-            <HeaderNetworkStatus onPress={goToNetworkSelect} />
-          ),
-          headerBackTitleVisible: false,
-          headerTitle: translate('screens/ReceiveScreen', 'Receive')
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={AddressControlScreen}
-          name='AddressControlScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/AddressControlScreen', 'Wallet Address')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerRightContainerStyle: tailwind('px-2 py-2'),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={GetDFIScreen}
-          name='GetDFIScreen'
-          options={{
-          ...screenOptions,
-          headerRight: () => (
-            <HeaderNetworkStatus onPress={goToNetworkSelect} />
-          ),
-          headerBackTitleVisible: false,
-          headerTitle: translate('screens/ReceiveScreen', 'Get DFI')
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={MarketplaceScreen}
-          name='MarketplaceScreen'
-          options={{
-          ...screenOptions,
-          headerRight: () => (
-            <HeaderNetworkStatus onPress={goToNetworkSelect} />
-          ),
-          headerBackTitleVisible: false,
-          headerTitle: translate('screens/MarketplaceScreen', 'Marketplace')
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={SendScreen}
-          name='Send'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/SendScreen', 'Send')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={SendConfirmationScreen}
-          name='SendConfirmationScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/SendConfirmationScreen', 'Confirm Send')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={TokenDetailScreen}
-          name='TokenDetail'
-          options={({ route }: { route: any }) => ({
-          headerTitle: () => {
-            const token = route?.params?.token
-            const Icon = getNativeIcon(token.displaySymbol)
-            return (
-              <HeaderTitle containerTestID={headerContainerTestId}>
-                <View style={tailwind('flex-row items-center')}>
-                  <Icon />
-
-                  <View style={tailwind('flex-col ml-2')}>
-                    <ThemedText style={tailwind('font-semibold')}>
-                      {token.displaySymbol}
-                    </ThemedText>
-
-                    <ConnectionStatus />
-                  </View>
-                </View>
-              </HeaderTitle>
-            )
-          },
-          headerBackTitleVisible: false
-        })}
-        />
-
-        <PortfolioStack.Screen
-          component={ConvertScreen}
-          name='Convert'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/ConvertScreen', 'Convert DFI')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={ConvertConfirmationScreen}
-          name='ConvertConfirmationScreen'
-          options={{
-          headerBackTitleVisible: false,
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/ConvertConfirmScreen', 'Confirm DFI Conversion')}
-              containerTestID={headerContainerTestId}
-            />
-          )
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={BarCodeScanner}
-          name='BarCodeScanner'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/ConvertScreen', 'Scan recipient QR')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={TokensVsUtxoScreen}
-          name='TokensVsUtxo'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/ConvertScreen', 'UTXO vs Token')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={NetworkDetails}
-          name='NetworkDetails'
-          options={{
-          headerTitle: translate('screens/NetworkDetails', 'Wallet Network'),
-          headerBackTitleVisible: false,
-          headerBackTestID: 'network_details_header_back'
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={AboutScreen}
-          name='AboutScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/AboutScreen', 'About')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={CompositeSwapScreen}
-          name='CompositeSwap'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/DexScreen', 'Swap tokens')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={ConfirmCompositeSwapScreen}
-          name='ConfirmCompositeSwapScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/DexScreen', 'Confirm swap')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={AddressBookScreen}
-          name='AddressBookScreen'
-          options={{
-          headerBackTitleVisible: false,
-          headerTitle: () => (
-            <></>
-          )
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={AddOrEditAddressBookScreen}
-          name='AddOrEditAddressBookScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/AddOrEditAddressBookScreen', 'Add New Address')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={AddLiquidityScreen}
-          name='AddLiquidity'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/DexScreen', 'Add Liquidity')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={ConfirmAddLiquidityScreen}
-          name='ConfirmAddLiquidity'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/DexScreen', 'Confirm Add Liquidity')}
-              containerTestID={headerContainerTestId}
-            />
-          )
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={WithdrawFutureSwapScreen}
-          name='WithdrawFutureSwapScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/WithdrawFutureSwapScreen', 'Withdraw from future swap')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={RemoveLiquidityScreen}
-          name='RemoveLiquidity'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/DexScreen', 'Remove Liquidity')}
-              containerTestID={headerContainerTestId}
-            />
-          )
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={FutureSwapScreen}
-          name='FutureSwapScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/FutureSwapScreen', 'Future Swap')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={ConfirmWithdrawFutureSwapScreen}
-          name='ConfirmWithdrawFutureSwapScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/ConfirmWithdrawFutureSwapScreen', 'Confirm withdrawal')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={RemoveLiquidityConfirmScreen}
-          name='RemoveLiquidityConfirmScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/DexScreen', 'Confirm Removal')}
-              containerTestID={headerContainerTestId}
-            />
-          )
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={TransactionsScreen}
-          name='TransactionsScreen'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/TransactionsScreen', 'Transactions')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-
-        <PortfolioStack.Screen
-          component={TransactionDetailScreen}
-          name='TransactionDetail'
-          options={{
-          headerTitle: () => (
-            <HeaderTitle
-              text={translate('screens/TransactionDetailScreen', 'Transaction')}
-              containerTestID={headerContainerTestId}
-            />
-          ),
-          headerBackTitleVisible: false
-        }}
-        />
-      </PortfolioStack.Navigator>
-    )
-  }
-
-  return (
-    <>
-      {isFeatureAvailable('portfolio_v2')
-        ? (<PortfolioStacksV2 />)
-        : (<PortfolioStacks />)}
-    </>
+      />
+    </PortfolioStack.Navigator>
   )
 }
